user_flow/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
awscli==1.33.44
boto3==1.34.162
botocore==1.34.162
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
docutils==0.16
exceptiongroup==1.2.2
fastapi==0.114.2
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-aws==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-postgres==0.0.12
langchain-text-splitters==0.3.0
langsmith==0.1.121
mangum==0.17.0
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pgvector==0.2.5
psycopg==3.2.2
psycopg-pool==3.2.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.11.1
